+++ test/git_changelog_generate.spec.js
@@ -293,7 +293,7 @@ describe('git_changelog_generate.js', function() {
 
         sections = changelog.organizeCommits(commits, sections);
 
-        expect(sections.fix.myModule.length).to.equal(10);
+        expect(sections.fix.mymodule.length).to.equal(10);
       });
 
     });