+++ CHANGELOG.md
@@ -19,9 +19,10 @@ _Git changelog is a utility tool for generating changelogs. It is free and opens
 
 ## Documentation
 
-- **Readme:** Unuseful commit
+- **readme:**
+  - Unuseful commit
   ([4373f472](https://github.com/rafinskipg/git-changelog/commit/4373f4726eedad6d450c8255f5e57036a3e5e223))
-- **readme:** fix link to the `.changelogrc` section
+  - fix link to the `.changelogrc` section
   ([2975171d](https://github.com/rafinskipg/git-changelog/commit/2975171d89e1823253399bbe87a184e9164e9799))